+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Tested with Ruby 3.1
 * Add `albumart` command
+* Add `readpicture` command
+
 ## 0.1.0
 
 * Rename `MPDClient` to `MPD::Client`